+++ packages/mp4-media-stream/vite.config.ts
@@ -24,7 +24,6 @@ export default defineConfig({
       formats: ["es"],
       name: "Shiguredo",
     },
-    manifest: true,
     minify: "esbuild",
     rolldownOptions: {
       output: {
@@ -52,7 +51,8 @@ export default defineConfig({
         },
       ],
     },
-    target: "es2023",
+    // ルート tsconfig.json の target と一致させる
+    target: "es2022",
   },
   plugins: [
     dts({