+++ packages/noise-suppression/vite.config.ts
@@ -23,14 +23,14 @@ export default defineConfig({
       formats: ["es"],
       name: "Shiguredo",
     },
-    manifest: true,
     minify: "esbuild",
     rolldownOptions: {
       output: {
         banner,
       },
     },
-    target: "es2023",
+    // ルート tsconfig.json の target と一致させる
+    target: "es2022",
   },
   plugins: [
     dts({